requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==2.2.2
django-filter==2.1.0
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
idna==2.8
Markdown==3.1.1
Pillow==6.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
